Refactor Race data calls
